Split getChangesFromHealthBot
